src/video_generator/requirements.txt
awscli
python-dotenv
urllib==1.26.9
bs4
elasticsearch==7.8.0
boto3==1.24.31
botocore==1.27.31